caloria_ai_pro_project/requirements.txt
aiogram==3.13.1
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
python-dotenv==1.0.1
aiosqlite==0.20.0
openai==1.51.0
